belajar class helpers

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\HeroController;
+use App\Http\Controllers\OrderController;
 use App\Http\Controllers\PromotionController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
@@ -24,15 +25,15 @@ Route::post('/login', [UserController::class, 'prosesLogin']);
 
 Route::get('/dashboard', function () {
     return view('pages.dashboard');
-})->middleware('auth');
+})->middleware(['auth', 'checkRole:admin']);
+Route::get('/hero', [HeroController::class, 'index'])->middleware(['auth', 'checkRole:admin']);
+Route::get('/order', [OrderController::class, 'index'])->middleware(['auth', 'checkRole:admin']);
+Route::get('/hero/create', [HeroController::class, 'create'])->middleware(['auth', 'checkRole:admin']);
+Route::post('/hero', [HeroController::class, 'store'])->middleware(['auth', 'checkRole:admin']);
+Route::get('/hero/{hero}/edit', [HeroController::class, 'edit'])->middleware(['auth', 'checkRole:admin']);
+Route::put('/hero/{hero}', [HeroController::class, 'update'])->middleware(['auth', 'checkRole:admin']);
+Route::delete('/hero/{hero}', [HeroController::class, 'destroy'])->middleware(['auth', 'checkRole:admin']);
 
-Route::get('/hero', [HeroController::class, 'index'])->middleware('auth');
-Route::get('/hero/create', [HeroController::class, 'create'])->middleware('auth');
-Route::post('/hero', [HeroController::class, 'store'])->middleware('auth');
-Route::get('/hero/{hero}/edit', [HeroController::class, 'edit'])->middleware('auth');
-Route::put('/hero/{hero}', [HeroController::class, 'update'])->middleware('auth');
-Route::delete('/hero/{hero}', [HeroController::class, 'destroy'])->middleware('auth');
+Route::resource('/promotion', PromotionController::class)->middleware(['auth', 'checkRole:admin']);
 
-Route::resource('/promotion', PromotionController::class)->middleware('auth');
-
-Route::get('/logout', [UserController::class, 'logout'])->name('logout')->middleware('auth');
+Route::get('/logout', [UserController::class, 'logout'])->name('logout')->middleware(['auth', 'checkRole:admin']);
